feat: extract reducer to custom hook
