+++ src/components/utils/copy-to-clipboard.tsx
@@ -1,3 +1,6 @@
+import { notificationService } from '../Snackbar';
+
 export default function copyToClipboard(text: string) {
   navigator.clipboard.writeText(text);
+  notificationService.openSuccessToast('Copied to clipboard');
 }
