+++ src/pages/startpage.tsx
@@ -10,7 +10,6 @@ import InfoSection from '@/components/home/InfoSection';
 import CtaSection from '@/components/annual-plans-for-affiliates/CtaSection';
 import PriceTable from '@/components/annual-plans-for-affiliates/components/PriceTable';
 import { checkout } from '@/lib/auth';
-import { useRouter } from 'next/router';
 import { CouponType } from '@/lib/types/types';
 
 const START_PAGE = 'STARTPAGE';
